requirements.txt
streamlit>=1.32.0
pymupdf
pandas>=2.2.0
yfinance>=0.2.36
rapidfuzz>=3.6.1
python-bidi>=0.4.2
arabic_reshaper>=3.0.0
plumber
plotly>=5.19.0
pdfplumber>=0.10.3
nltk>=3.8.1
matplotlib>=3.8.3
numpy>=1.26.4
cryptography>=42.0.5
cffi>=1.16.0
joblib>=1.3.2
regex>=2023.12.25
tqdm>=4.66.2
beautifulsoup4>=4.12.3
requests>=2.31.0
python-dateutil>=2.8.2